+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 A preview of major changes can be found in the Wiki ([Latest Changes](https://github.com/kcapp/frontend/wiki/Latest_Changes))
+
 ## [2.6.0] - TBD
+#### Feature
+- New query parameter `officeId=<id>` for `/obs` view, to only forward on matches in a given office
+
 #### Fixed
 - Bug where vocal announcement for Cricket would say "No score" instead of marks